+++ lib/components/player_info_screen/gadget_list.dart
@@ -1,3 +1,5 @@
+import 'package:battlefield_2042_state/components/basic/constraints_modal_bottom_sheet.dart';
+import 'package:battlefield_2042_state/components/basic/info_list_item_content.dart';
 import 'package:battlefield_2042_state/components/basic/player_detail_info_list.dart';
 import 'package:battlefield_2042_state/model/player_info_ensemble.dart';
 import 'package:battlefield_2042_state/model/player_info_model.dart';
@@ -5,12 +7,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 
-import 'basic/constraints_modal_bottom_sheet.dart';
-import 'basic/info_list_item_content.dart';
-
 enum DataType {
   kpm('killsPerMinute'),
-  secondsPlayed('timePlayed'),
+  uses('uses'),
   destroyCount('destroyCount');
 
   const DataType(this.value);
@@ -18,54 +17,42 @@ enum DataType {
   final String value;
 }
 
-class VehicleList extends StatefulWidget {
-  const VehicleList({super.key});
+class GadgetList extends StatefulWidget {
+  const GadgetList({super.key});
 
   @override
-  VehicleListState createState() => VehicleListState();
+  GadgetListState createState() => GadgetListState();
 }
 
-class VehicleListState extends State<VehicleList> {
+class GadgetListState extends State<GadgetList> {
   String dataTypeValue = 'killsPerMinute';
 
-  void showVehicleDetails(BuildContext context, VehicleInfoEnsemble vehicle) {
-    final List<InfoListItemContent> vehicleDetailList = [
+  void showVehicleDetails(BuildContext context, GadgetInfoEnsemble gadget) {
+    final List<InfoListItemContent> gadgetDetailList = [
       InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.killsTitle,
+          keyName: AppLocalizations.of(context)!.kills,
           showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(vehicle.kills)),
+          AppLocalizations.of(context)!.universalIntDisplay(gadget.kills)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.kpmTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalDoubleDisplay(vehicle.KPM)),
+          showValueString:
+          AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.vehiclesDestroyed,
           showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.killedVehicle)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.roadKillsTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.roadKills)),
+              .universalIntDisplay(gadget.killedVehicle)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.damage,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.damage)),
+          showValueString:
+          AppLocalizations.of(context)!.universalIntDisplay(gadget.damage)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.multiKillsTitle,
           showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.multiKills)),
+              .universalIntDisplay(gadget.multiKills)),
       InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.driverAssistsTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.driverAssists)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.passengerAssistsTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.passengerAssists)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.distanceTraveledTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(vehicle.distanceTraveled)),
+          keyName: AppLocalizations.of(context)!.usesTitle,
+          showValueString:
+          AppLocalizations.of(context)!.universalIntDisplay(gadget.used)),
     ];
 
     ConstraintsModalBottomSheet.showConstraintsModalBottomSheet(
@@ -77,34 +64,28 @@ class VehicleListState extends State<VehicleList> {
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              Column(
+              Row(
                   mainAxisAlignment: MainAxisAlignment.center,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
                     FittedBox(
                       fit: BoxFit.scaleDown,
-                      child: Text(vehicle.vehicleName,
+                      child: Text(gadget.gadgetName,
                           style: Theme.of(context).textTheme.titleLarge),
                     ),
-                    const Padding(padding: EdgeInsets.only(left: 16)),
-                    Badge(
-                        backgroundColor: Theme.of(context).colorScheme.primary,
-                        label: Text(AppLocalizations.of(context)!
-                            .playedTime(vehicle.playedTime)))
                   ]),
               Expanded(
                   child: ListView.builder(
-                shrinkWrap: true,
-                prototypeItem: const InfoListItem(
-                    keyName: 'null', showValueString: 'null'),
-                itemCount: vehicleDetailList.length,
-                itemBuilder: (context, index) {
-                  return InfoListItem(
-                      keyName: vehicleDetailList[index].keyName,
-                      showValueString:
-                          vehicleDetailList[index].showValueString);
-                },
-              ))
+                    shrinkWrap: true,
+                    prototypeItem: const InfoListItem(
+                        keyName: 'null', showValueString: 'null'),
+                    itemCount: gadgetDetailList.length,
+                    itemBuilder: (context, index) {
+                      return InfoListItem(
+                          keyName: gadgetDetailList[index].keyName,
+                          showValueString: gadgetDetailList[index].showValueString);
+                    },
+                  ))
             ],
           ),
         ));
@@ -120,11 +101,12 @@ class VehicleListState extends State<VehicleList> {
             return ListTile(
               title: Text(
                 AppLocalizations.of(context)!
-                    .vehicleDataType(DataType.values[index].value),
+                    .gadgetDataType(DataType.values[index].value),
                 style: Theme.of(context).textTheme.titleMedium,
               ),
               shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(19)),
+                borderRadius: BorderRadius.circular(19),
+              ),
               tileColor: dataTypeValue == DataType.values[index].value
                   ? Theme.of(context).colorScheme.secondaryContainer
                   : null,
@@ -142,14 +124,14 @@ class VehicleListState extends State<VehicleList> {
   @override
   Widget build(BuildContext context) {
     return Consumer<PlayerInfoModel>(builder: (context, playerInfo, child) {
-      final vehicleList = playerInfo.playerInfoEnsemble.vehicles;
-      vehicleList.sort((a, b) => (b.kills - a.kills));
+      final gadgetList = playerInfo.playerInfoEnsemble.gadgets;
+      gadgetList.sort((a, b) => (b.kills - a.kills));
 
       return TouchableList(
           listTitle: [
             Expanded(
                 flex: 2,
-                child: Text(AppLocalizations.of(context)!.vehicleNameTitle,
+                child: Text(AppLocalizations.of(context)!.gadgetNameTitle,
                     softWrap: true,
                     textAlign: TextAlign.left,
                     style: Theme.of(context).textTheme.bodyLarge)),
@@ -169,7 +151,7 @@ class VehicleListState extends State<VehicleList> {
                       children: [
                         Text(
                           AppLocalizations.of(context)!
-                              .vehicleDataType(dataTypeValue),
+                              .gadgetDataType(dataTypeValue),
                           textAlign: TextAlign.right,
                           style: Theme.of(context).textTheme.bodyLarge,
                           overflow: TextOverflow.ellipsis,
@@ -180,67 +162,58 @@ class VehicleListState extends State<VehicleList> {
                           size: 16,
                         )
                       ],
-                    ))),
+                    )))
           ],
           listChild: ListView.builder(
               shrinkWrap: true,
-              prototypeItem: VehicleListItem(
-                vehicle: VehicleInfoEnsemble(
-                  'null',
-                  'null',
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                  0,
-                ),
-                dataTypeValue: dataTypeValue,
-              ),
+              prototypeItem: GadgetListItem(
+                  dataTypeValue: dataTypeValue,
+                  gadget: GadgetInfoEnsemble(
+                    'null',
+                    'null',
+                    0,
+                    0,
+                    0,
+                    0,
+                    0,
+                    0,
+                  )),
               itemCount: playerInfo.playerInfoEnsemble.vehicles.length,
               itemBuilder: (context, index) {
-                return VehicleListItem(
-                  vehicle: playerInfo.playerInfoEnsemble.vehicles[index],
+                return GadgetListItem(
+                  dataTypeValue: dataTypeValue,
+                  gadget: playerInfo.playerInfoEnsemble.gadgets[index],
                   onTap: () => {
                     showVehicleDetails(
-                        context, playerInfo.playerInfoEnsemble.vehicles[index])
+                        context, playerInfo.playerInfoEnsemble.gadgets[index])
                   },
-                  dataTypeValue: dataTypeValue,
                 );
               }));
     });
   }
 }
 
-class VehicleListItem extends StatelessWidget {
-  final VehicleInfoEnsemble vehicle;
+class GadgetListItem extends StatelessWidget {
+  final GadgetInfoEnsemble gadget;
   final String dataTypeValue;
   final Function? onTap;
 
-  const VehicleListItem(
-      {super.key,
-      required this.vehicle,
-      required this.dataTypeValue,
-      this.onTap});
+  const GadgetListItem({super.key,
+    required this.gadget,
+    required this.dataTypeValue,
+    this.onTap});
 
-  String filterVehicleDataByDataTypeValue(
-      BuildContext context, VehicleInfoEnsemble vehicle, String value) {
+  String filterGadgetDataByDataTypeValue(BuildContext context, GadgetInfoEnsemble gadget, String value) {
     switch (value) {
       case 'killsPerMinute':
-        return AppLocalizations.of(context)!
-            .universalDoubleDisplay(vehicle.KPM);
-      case 'timePlayed':
-        return AppLocalizations.of(context)!.playedTime(vehicle.playedTime);
+        return AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM);
+      case 'uses':
+        return AppLocalizations.of(context)!.universalIntDisplay(gadget.used);
       case 'destroyCount':
         return AppLocalizations.of(context)!
-            .universalIntDisplay(vehicle.killedVehicle);
+            .universalIntDisplay(gadget.killedVehicle);
       default:
-        return AppLocalizations.of(context)!
-            .universalDoubleDisplay(vehicle.KPM);
+        return AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM);
     }
   }
 
@@ -251,14 +224,14 @@ class VehicleListItem extends StatelessWidget {
           flex: 2,
           child: SingleChildScrollView(
             scrollDirection: Axis.horizontal,
-            child: Text(vehicle.vehicleName,
+            child: Text(gadget.gadgetName,
                 textAlign: TextAlign.left,
                 style: Theme.of(context).textTheme.bodyMedium),
           )),
       Expanded(
           flex: 1,
           child: Text(
-            AppLocalizations.of(context)!.universalIntDisplay(vehicle.kills),
+            AppLocalizations.of(context)!.universalIntDisplay(gadget.kills),
             textAlign: TextAlign.center,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
@@ -269,14 +242,14 @@ class VehicleListItem extends StatelessWidget {
       Expanded(
           flex: 1,
           child: Text(
-            filterVehicleDataByDataTypeValue(context, vehicle, dataTypeValue),
+            filterGadgetDataByDataTypeValue(context, gadget, dataTypeValue),
             textAlign: TextAlign.right,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
               fontSize: Theme.of(context).textTheme.bodyMedium?.fontSize,
               color: Theme.of(context).colorScheme.primary,
             ),
-          )),
+          ))
     ];
 
     return TouchableListItem(
