+++ lib/components/player_info_screen/map_list.dart
@@ -6,66 +6,30 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 
-import 'basic/constraints_modal_bottom_sheet.dart';
-import 'basic/info_list_item_content.dart';
+import '../basic/constraints_modal_bottom_sheet.dart';
+import '../basic/info_list_item_content.dart';
 
-class GameModeList extends StatelessWidget {
-  const GameModeList({super.key});
+class MapList extends StatelessWidget {
+  const MapList({super.key});
 
-  void showVehicleDetails(BuildContext context, GameModeInfoEnsemble gamemode) {
-    final List<InfoListItemContent> gameModeDetailList = [
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.kills,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.kills)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.kpmTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalDoubleDisplay(gamemode.KPM)),
+  void showVehicleDetails(BuildContext context, MapInfoEnsemble mapElement) {
+    final List<InfoListItemContent> mapDetailList = [
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.matches,
           showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.playedMatches)),
+              .universalIntDisplay(mapElement.playedMatches)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.wins,
-          showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(gamemode.win)),
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(mapElement.win)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.losses,
-          showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(gamemode.lose)),
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(mapElement.lose)),
       InfoListItemContent(
         keyName: AppLocalizations.of(context)!.winRate,
-        showValueString: gamemode.winRate,
+        showValueString: mapElement.winRate,
       ),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectTimeTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.objectTime)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.sectorDefendsTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.sectorDefend)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectDefendsTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.objectDefend)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectTakenTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.objectCapture)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectArmedTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.boomPlant)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectDisarmedTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.boomDefuse)),
-      InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.objectDestroyedTitle,
-          showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gamemode.boomDestroy)),
     ];
 
     ConstraintsModalBottomSheet.showConstraintsModalBottomSheet(
@@ -86,27 +50,26 @@ class GameModeList extends StatelessWidget {
                       child: Text(
                           Translator.appLocalizationsTranslate(
                               AppLocalizations.of(context)!
-                                  .gameModeName(gamemode.modeName),
-                              gamemode.modeName),
+                                  .mapName(mapElement.mapName),
+                              mapElement.mapName),
                           style: Theme.of(context).textTheme.titleLarge),
                     ),
                     const Padding(padding: EdgeInsets.only(left: 8)),
                     Badge(
                         backgroundColor: Theme.of(context).colorScheme.primary,
                         label: Text(AppLocalizations.of(context)!
-                            .playedTime(gamemode.playedTime)))
+                            .playedTime(mapElement.playedTime)))
                   ]),
               Expanded(
                   child: ListView.builder(
                 shrinkWrap: true,
                 prototypeItem: const InfoListItem(
                     keyName: 'null', showValueString: 'null'),
-                itemCount: gameModeDetailList.length,
+                itemCount: mapDetailList.length,
                 itemBuilder: (context, index) {
                   return InfoListItem(
-                      keyName: gameModeDetailList[index].keyName,
-                      showValueString:
-                          gameModeDetailList[index].showValueString);
+                      keyName: mapDetailList[index].keyName,
+                      showValueString: mapDetailList[index].showValueString);
                 },
               ))
             ],
@@ -117,14 +80,14 @@ class GameModeList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Consumer<PlayerInfoModel>(builder: (context, playerInfo, child) {
-      final gameModeList = playerInfo.playerInfoEnsemble.gameModes;
-      gameModeList.sort((a, b) => (b.playedMatches - a.playedMatches));
+      final mapList = playerInfo.playerInfoEnsemble.maps;
+      mapList.sort((a, b) => (b.playedMatches - a.playedMatches));
 
       return TouchableList(
           listTitle: [
             Expanded(
                 flex: 2,
-                child: Text(AppLocalizations.of(context)!.gamemodeNameTitle,
+                child: Text(AppLocalizations.of(context)!.mapNameTitle,
                     softWrap: true,
                     textAlign: TextAlign.left,
                     style: Theme.of(context).textTheme.bodyLarge)),
@@ -141,32 +104,23 @@ class GameModeList extends StatelessWidget {
           ],
           listChild: ListView.builder(
               shrinkWrap: true,
-              prototypeItem: GameModeListItem(
-                  gamemode: GameModeInfoEnsemble(
+              prototypeItem: MapListItem(
+                  mapElement: MapInfoEnsemble(
                 'null',
                 'null',
                 0,
                 0,
-                0,
-                0,
-                0,
                 'null',
-                0,
-                0,
-                0,
-                0,
-                0,
-                0,
                 0,
                 0,
               )),
-              itemCount: playerInfo.playerInfoEnsemble.gameModes.length,
+              itemCount: playerInfo.playerInfoEnsemble.maps.length,
               itemBuilder: (context, index) {
-                return GameModeListItem(
-                  gamemode: playerInfo.playerInfoEnsemble.gameModes[index],
+                return MapListItem(
+                  mapElement: playerInfo.playerInfoEnsemble.maps[index],
                   onTap: () => {
                     showVehicleDetails(
-                        context, playerInfo.playerInfoEnsemble.gameModes[index])
+                        context, playerInfo.playerInfoEnsemble.maps[index])
                   },
                 );
               }));
@@ -174,11 +128,11 @@ class GameModeList extends StatelessWidget {
   }
 }
 
-class GameModeListItem extends StatelessWidget {
-  final GameModeInfoEnsemble gamemode;
+class MapListItem extends StatelessWidget {
+  final MapInfoEnsemble mapElement;
   final Function? onTap;
 
-  const GameModeListItem({super.key, required this.gamemode, this.onTap});
+  const MapListItem({super.key, required this.mapElement, this.onTap});
 
   @override
   Widget build(BuildContext context) {
@@ -189,16 +143,16 @@ class GameModeListItem extends StatelessWidget {
             scrollDirection: Axis.horizontal,
             child: Text(
                 Translator.appLocalizationsTranslate(
-                    AppLocalizations.of(context)!
-                        .gameModeName(gamemode.modeName),
-                    gamemode.modeName),
+                    AppLocalizations.of(context)!.mapName(mapElement.mapName),
+                    mapElement.mapName),
+                textAlign: TextAlign.left,
                 style: Theme.of(context).textTheme.bodyMedium),
           )),
       Expanded(
           flex: 1,
           child: Text(
             AppLocalizations.of(context)!
-                .universalIntDisplay(gamemode.playedMatches),
+                .universalIntDisplay(mapElement.playedMatches),
             textAlign: TextAlign.center,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
@@ -209,7 +163,7 @@ class GameModeListItem extends StatelessWidget {
       Expanded(
           flex: 1,
           child: Text(
-            gamemode.winRate,
+            mapElement.winRate,
             textAlign: TextAlign.right,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
