+++ lib/components/player_info_screen/overview_list.dart
@@ -1,9 +1,8 @@
 import 'package:battlefield_2042_state/components/basic/info_list_item_content.dart';
+import 'package:battlefield_2042_state/model/player_info_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
-
-import '../model/player_info_model.dart';
 
 class OverviewList extends StatelessWidget {
   const OverviewList({super.key});
