+++ lib/components/player_info_screen/vehicle_list.dart
@@ -5,12 +5,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 
-import 'basic/constraints_modal_bottom_sheet.dart';
-import 'basic/info_list_item_content.dart';
+import '../basic/constraints_modal_bottom_sheet.dart';
+import '../basic/info_list_item_content.dart';
 
 enum DataType {
   kpm('killsPerMinute'),
-  uses('uses'),
+  secondsPlayed('timePlayed'),
   destroyCount('destroyCount');
 
   const DataType(this.value);
@@ -18,42 +18,54 @@ enum DataType {
   final String value;
 }
 
-class GadgetList extends StatefulWidget {
-  const GadgetList({super.key});
+class VehicleList extends StatefulWidget {
+  const VehicleList({super.key});
 
   @override
-  GadgetListState createState() => GadgetListState();
+  VehicleListState createState() => VehicleListState();
 }
 
-class GadgetListState extends State<GadgetList> {
+class VehicleListState extends State<VehicleList> {
   String dataTypeValue = 'killsPerMinute';
 
-  void showVehicleDetails(BuildContext context, GadgetInfoEnsemble gadget) {
-    final List<InfoListItemContent> gadgetDetailList = [
+  void showVehicleDetails(BuildContext context, VehicleInfoEnsemble vehicle) {
+    final List<InfoListItemContent> vehicleDetailList = [
       InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.kills,
+          keyName: AppLocalizations.of(context)!.killsTitle,
           showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(gadget.kills)),
+              AppLocalizations.of(context)!.universalIntDisplay(vehicle.kills)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.kpmTitle,
-          showValueString:
-              AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM)),
+          showValueString: AppLocalizations.of(context)!
+              .universalDoubleDisplay(vehicle.KPM)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.vehiclesDestroyed,
           showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gadget.killedVehicle)),
+              .universalIntDisplay(vehicle.killedVehicle)),
+      InfoListItemContent(
+          keyName: AppLocalizations.of(context)!.roadKillsTitle,
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(vehicle.roadKills)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.damage,
-          showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(gadget.damage)),
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(vehicle.damage)),
       InfoListItemContent(
           keyName: AppLocalizations.of(context)!.multiKillsTitle,
           showValueString: AppLocalizations.of(context)!
-              .universalIntDisplay(gadget.multiKills)),
+              .universalIntDisplay(vehicle.multiKills)),
       InfoListItemContent(
-          keyName: AppLocalizations.of(context)!.usesTitle,
-          showValueString:
-              AppLocalizations.of(context)!.universalIntDisplay(gadget.used)),
+          keyName: AppLocalizations.of(context)!.driverAssistsTitle,
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(vehicle.driverAssists)),
+      InfoListItemContent(
+          keyName: AppLocalizations.of(context)!.passengerAssistsTitle,
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(vehicle.passengerAssists)),
+      InfoListItemContent(
+          keyName: AppLocalizations.of(context)!.distanceTraveledTitle,
+          showValueString: AppLocalizations.of(context)!
+              .universalIntDisplay(vehicle.distanceTraveled)),
     ];
 
     ConstraintsModalBottomSheet.showConstraintsModalBottomSheet(
@@ -65,26 +77,32 @@ class GadgetListState extends State<GadgetList> {
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              Row(
+              Column(
                   mainAxisAlignment: MainAxisAlignment.center,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
                     FittedBox(
                       fit: BoxFit.scaleDown,
-                      child: Text(gadget.gadgetName,
+                      child: Text(vehicle.vehicleName,
                           style: Theme.of(context).textTheme.titleLarge),
                     ),
+                    const Padding(padding: EdgeInsets.only(left: 16)),
+                    Badge(
+                        backgroundColor: Theme.of(context).colorScheme.primary,
+                        label: Text(AppLocalizations.of(context)!
+                            .playedTime(vehicle.playedTime)))
                   ]),
               Expanded(
                   child: ListView.builder(
                 shrinkWrap: true,
                 prototypeItem: const InfoListItem(
                     keyName: 'null', showValueString: 'null'),
-                itemCount: gadgetDetailList.length,
+                itemCount: vehicleDetailList.length,
                 itemBuilder: (context, index) {
                   return InfoListItem(
-                      keyName: gadgetDetailList[index].keyName,
-                      showValueString: gadgetDetailList[index].showValueString);
+                      keyName: vehicleDetailList[index].keyName,
+                      showValueString:
+                          vehicleDetailList[index].showValueString);
                 },
               ))
             ],
@@ -102,12 +120,11 @@ class GadgetListState extends State<GadgetList> {
             return ListTile(
               title: Text(
                 AppLocalizations.of(context)!
-                    .gadgetDataType(DataType.values[index].value),
+                    .vehicleDataType(DataType.values[index].value),
                 style: Theme.of(context).textTheme.titleMedium,
               ),
               shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(19),
-              ),
+                  borderRadius: BorderRadius.circular(19)),
               tileColor: dataTypeValue == DataType.values[index].value
                   ? Theme.of(context).colorScheme.secondaryContainer
                   : null,
@@ -125,14 +142,14 @@ class GadgetListState extends State<GadgetList> {
   @override
   Widget build(BuildContext context) {
     return Consumer<PlayerInfoModel>(builder: (context, playerInfo, child) {
-      final gadgetList = playerInfo.playerInfoEnsemble.gadgets;
-      gadgetList.sort((a, b) => (b.kills - a.kills));
+      final vehicleList = playerInfo.playerInfoEnsemble.vehicles;
+      vehicleList.sort((a, b) => (b.kills - a.kills));
 
       return TouchableList(
           listTitle: [
             Expanded(
                 flex: 2,
-                child: Text(AppLocalizations.of(context)!.gadgetNameTitle,
+                child: Text(AppLocalizations.of(context)!.vehicleNameTitle,
                     softWrap: true,
                     textAlign: TextAlign.left,
                     style: Theme.of(context).textTheme.bodyLarge)),
@@ -152,7 +169,7 @@ class GadgetListState extends State<GadgetList> {
                       children: [
                         Text(
                           AppLocalizations.of(context)!
-                              .gadgetDataType(dataTypeValue),
+                              .vehicleDataType(dataTypeValue),
                           textAlign: TextAlign.right,
                           style: Theme.of(context).textTheme.bodyLarge,
                           overflow: TextOverflow.ellipsis,
@@ -163,60 +180,67 @@ class GadgetListState extends State<GadgetList> {
                           size: 16,
                         )
                       ],
-                    )))
+                    ))),
           ],
           listChild: ListView.builder(
               shrinkWrap: true,
-              prototypeItem: GadgetListItem(
-                  dataTypeValue: dataTypeValue,
-                  gadget: GadgetInfoEnsemble(
-                    'null',
-                    'null',
-                    0,
-                    0,
-                    0,
-                    0,
-                    0,
-                    0,
-                  )),
+              prototypeItem: VehicleListItem(
+                vehicle: VehicleInfoEnsemble(
+                  'null',
+                  'null',
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                  0,
+                ),
+                dataTypeValue: dataTypeValue,
+              ),
               itemCount: playerInfo.playerInfoEnsemble.vehicles.length,
               itemBuilder: (context, index) {
-                return GadgetListItem(
-                  dataTypeValue: dataTypeValue,
-                  gadget: playerInfo.playerInfoEnsemble.gadgets[index],
+                return VehicleListItem(
+                  vehicle: playerInfo.playerInfoEnsemble.vehicles[index],
                   onTap: () => {
                     showVehicleDetails(
-                        context, playerInfo.playerInfoEnsemble.gadgets[index])
+                        context, playerInfo.playerInfoEnsemble.vehicles[index])
                   },
+                  dataTypeValue: dataTypeValue,
                 );
               }));
     });
   }
 }
 
-class GadgetListItem extends StatelessWidget {
-  final GadgetInfoEnsemble gadget;
+class VehicleListItem extends StatelessWidget {
+  final VehicleInfoEnsemble vehicle;
   final String dataTypeValue;
   final Function? onTap;
 
-  const GadgetListItem(
+  const VehicleListItem(
       {super.key,
-      required this.gadget,
+      required this.vehicle,
       required this.dataTypeValue,
       this.onTap});
 
-  String filterGadgetDataByDataTypeValue(
-      BuildContext context, GadgetInfoEnsemble gadget, String value) {
+  String filterVehicleDataByDataTypeValue(
+      BuildContext context, VehicleInfoEnsemble vehicle, String value) {
     switch (value) {
       case 'killsPerMinute':
-        return AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM);
-      case 'uses':
-        return AppLocalizations.of(context)!.universalIntDisplay(gadget.used);
+        return AppLocalizations.of(context)!
+            .universalDoubleDisplay(vehicle.KPM);
+      case 'timePlayed':
+        return AppLocalizations.of(context)!.playedTime(vehicle.playedTime);
       case 'destroyCount':
         return AppLocalizations.of(context)!
-            .universalIntDisplay(gadget.killedVehicle);
+            .universalIntDisplay(vehicle.killedVehicle);
       default:
-        return AppLocalizations.of(context)!.universalDoubleDisplay(gadget.KPM);
+        return AppLocalizations.of(context)!
+            .universalDoubleDisplay(vehicle.KPM);
     }
   }
 
@@ -227,14 +251,14 @@ class GadgetListItem extends StatelessWidget {
           flex: 2,
           child: SingleChildScrollView(
             scrollDirection: Axis.horizontal,
-            child: Text(gadget.gadgetName,
+            child: Text(vehicle.vehicleName,
                 textAlign: TextAlign.left,
                 style: Theme.of(context).textTheme.bodyMedium),
           )),
       Expanded(
           flex: 1,
           child: Text(
-            AppLocalizations.of(context)!.universalIntDisplay(gadget.kills),
+            AppLocalizations.of(context)!.universalIntDisplay(vehicle.kills),
             textAlign: TextAlign.center,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
@@ -245,14 +269,14 @@ class GadgetListItem extends StatelessWidget {
       Expanded(
           flex: 1,
           child: Text(
-            filterGadgetDataByDataTypeValue(context, gadget, dataTypeValue),
+            filterVehicleDataByDataTypeValue(context, vehicle, dataTypeValue),
             textAlign: TextAlign.right,
             style: TextStyle(
               fontWeight: Theme.of(context).textTheme.bodyMedium?.fontWeight,
               fontSize: Theme.of(context).textTheme.bodyMedium?.fontSize,
               color: Theme.of(context).colorScheme.primary,
             ),
-          ))
+          )),
     ];
 
     return TouchableListItem(
