+++ lib/components/player_info_screen/weapon_list.dart
