+++ lib/screen/player_info_screen.dart
@@ -4,15 +4,15 @@ import 'package:battlefield_2042_state/api/gametools/player_feslid.dart';
 import 'package:battlefield_2042_state/components/basic/constraints_modal_bottom_sheet.dart';
 import 'package:battlefield_2042_state/components/basic/custom_snackbar.dart';
 import 'package:battlefield_2042_state/components/basic/info_list_item_content.dart';
-import 'package:battlefield_2042_state/components/classes_list.dart';
-import 'package:battlefield_2042_state/components/gadget_list.dart';
-import 'package:battlefield_2042_state/components/gamemode_list.dart';
-import 'package:battlefield_2042_state/components/key_info_view.dart';
-import 'package:battlefield_2042_state/components/map_list.dart';
-import 'package:battlefield_2042_state/components/overview_list.dart';
-import 'package:battlefield_2042_state/components/share_player_stats.dart';
-import 'package:battlefield_2042_state/components/vehicle_list.dart';
-import 'package:battlefield_2042_state/components/weapon_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/classes_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/gadget_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/gamemode_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/key_info_view.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/map_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/overview_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/share_player_stats.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/vehicle_list.dart';
+import 'package:battlefield_2042_state/components/player_info_screen/weapon_list.dart';
 import 'package:battlefield_2042_state/model/player_info_ensemble.dart';
 import 'package:battlefield_2042_state/model/player_info_model.dart';
 import 'package:battlefield_2042_state/model/player_info_tab.dart';
